+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Move `RevitParameters` to `DesignAutomationController`
 - Add `TargetFramework` in `DesignAutomationController`
 - Update `App` to use abstract `DesignApplication` to load correct assembly in the bundle.
-- Update `ricaun.NUnit` to version `1.4.0-beta.2`, `CaseSource` and `ExportedTypes` support.
+- Update `ricaun.NUnit` to version `1.4.0-rc.1`, `CaseSource` and `ExportedTypes` support.
 - Add `ExternalServer` to force `DesignAutomationReadyEvent` event to trigger inside the server and execute with a valid `ActiveAddInId`.
 ### Console
 - Update `ricaun.Autodesk.Forge.Oss.DesignAutomation` to `2.0.0`